Refactored project to use Maven project structure and JNI loading utility.
