+++ .gitignore
@@ -1,0 +1,20 @@
+# IntelliJ exclusions
+.idea/
+*.iml
+*.ipr
+/out/
+
+# Java exclusions
+*.class
+*.jar
+
+# Maven exclusions
+target/
+pom.xml.tag
+pom.xml.releaseBackup
+pom.xml.versionsBackup
+pom.xml.next
+release.properties
+dependency-reduced-pom.xml
+buildNumber.properties
+.mvn/timing.properties
