+++ java-hashes/src/main/java/nayuki/nativehash/Md2Java.java
