+++ java-hashes/src/main/java/nayuki/nativehash/Md4Java.java
