+++ java-hashes/src/main/java/nayuki/nativehash/Md5Java.java
