+++ java-hashes/src/main/java/nayuki/nativehash/Ripemd128Java.java
