+++ java-hashes/src/main/java/nayuki/nativehash/Sha1Java.java
