+++ java-hashes/src/main/java/nayuki/nativehash/Tiger2Java.java
