+++ java-hashes/src/main/java/nayuki/nativehash/TigerJava.java
