+++ java-hashes/src/main/java/nayuki/nativehash/WhirlpoolJava.java
