+++ java-hashes/src/test/java/nayuki/nativehash/HashTest.java
