+++ java-hashes/src/test/java/nayuki/nativehash/Ripemd128Test.java
