+++ java-hashes/src/test/java/nayuki/nativehash/Sha1Test.java
