+++ java-hashes/src/test/java/nayuki/nativehash/Tiger2Test.java
