+++ java-hashes/src/test/java/nayuki/nativehash/WhirlpoolTest.java
