+++ native-hashes-benchmark/src/main/java/nayuki/nativehash/BenchmarkHashes.java
