+++ native-hashes-benchmark/src/main/java/sha1sum.java
