+++ native-hashes/src/main/java/nayuki/nativehash/BlockHasher.java
