+++ native-hashes/src/main/java/nayuki/nativehash/Md4.java
@@ -10,13 +10,13 @@ package nayuki.nativehash;
 import java.util.Arrays;
 
 
-public class Md5 extends BlockHasher {
+public class Md4 extends NativeBlockHasher {
 	
 	protected int[] state;
 	
 	
 	
-	public Md5() {
+	public Md4() {
 		super(64);
 		state = new int[]{0x67452301, 0xEFCDAB89, 0x98BADCFE, 0x10325476};
 	}
