+++ native-hashes/src/main/java/nayuki/nativehash/Ripemd256.java
@@ -10,17 +10,15 @@ package nayuki.nativehash;
 import java.util.Arrays;
 
 
-public class Tiger extends BlockHasher {
+public class Ripemd256 extends NativeBlockHasher {
 	
-	protected long[] state;
-	protected byte padding;
-	
+	protected int[] state;
 	
 	
-	public Tiger() {
+	
+	public Ripemd256() {
 		super(64);
-		state = new long[]{0x0123456789ABCDEFL, 0xFEDCBA9876543210L, 0xF096A5B4C3B2E187L};
-		padding = 0x01;
+		state = new int[]{0x67452301, 0xEFCDAB89, 0x98BADCFE, 0x10325476, 0x76543210, 0xFEDCBA98, 0x89ABCDEF, 0x01234567};
 	}
 	
 	
@@ -32,7 +30,7 @@ public class Tiger extends BlockHasher {
 	
 	
 	protected byte[] getHashDestructively() {
-		block[blockFilled] = padding;
+		block[blockFilled] = (byte)0x80;
 		blockFilled++;
 		Arrays.fill(block, blockFilled, block.length, (byte)0);
 		if (blockFilled + 8 > block.length) {
@@ -44,14 +42,14 @@ public class Tiger extends BlockHasher {
 			block[block.length - 8 + i] = (byte)(length >>> (i * 8));
 		compress(block, 0, block.length);
 		
-		byte[] result = new byte[state.length * 8];
+		byte[] result = new byte[state.length * 4];
 		for (int i = 0; i < result.length; i++)
-			result[i] = (byte)(state[i / 8] >>> (i % 8 * 8));
+			result[i] = (byte)(state[i / 4] >>> (i % 4 * 8));
 		return result;
 	}
 	
 	
-	private static native boolean compress(long[] state, byte[] msg, int off, int len);
+	private static native boolean compress(int[] state, byte[] msg, int off, int len);
 	
 	
 	static {
