+++ native-hashes/src/main/java/nayuki/nativehash/Tiger.java
@@ -10,15 +10,17 @@ package nayuki.nativehash;
 import java.util.Arrays;
 
 
-public class Whirlpool extends BlockHasher {
+public class Tiger extends NativeBlockHasher {
 	
-	protected byte[] state;
+	protected long[] state;
+	protected byte padding;
 	
 	
 	
-	public Whirlpool() {
+	public Tiger() {
 		super(64);
-		state = new byte[64];
+		state = new long[]{0x0123456789ABCDEFL, 0xFEDCBA9876543210L, 0xF096A5B4C3B2E187L};
+		padding = 0x01;
 	}
 	
 	
@@ -30,23 +32,26 @@ public class Whirlpool extends BlockHasher {
 	
 	
 	protected byte[] getHashDestructively() {
-		block[blockFilled] = (byte)0x80;
+		block[blockFilled] = padding;
 		blockFilled++;
 		Arrays.fill(block, blockFilled, block.length, (byte)0);
-		if (blockFilled + 32 > block.length) {
+		if (blockFilled + 8 > block.length) {
 			compress(block, 0, block.length);
 			Arrays.fill(block, (byte)0);
 		}
 		length = length << 3;
 		for (int i = 0; i < 8; i++)
-			block[block.length - 1 - i] = (byte)(length >>> (i * 8));
+			block[block.length - 8 + i] = (byte)(length >>> (i * 8));
 		compress(block, 0, block.length);
 		
-		return state;
+		byte[] result = new byte[state.length * 8];
+		for (int i = 0; i < result.length; i++)
+			result[i] = (byte)(state[i / 8] >>> (i % 8 * 8));
+		return result;
 	}
 	
 	
-	private static native boolean compress(byte[] state, byte[] msg, int off, int len);
+	private static native boolean compress(long[] state, byte[] msg, int off, int len);
 	
 	
 	static {
