+++ native-hashes/src/main/java/nayuki/nativehash/Tiger2.java
